Unit test view
