+++ apilos_settings/tests/test_views.py
@@ -1,5 +1,9 @@
+import os
+
+from django.core.files.uploadedfile import SimpleUploadedFile
 from django.test import TestCase
 from django.urls import reverse
+from rest_framework import status
 
 from bailleurs.models import Bailleur
 from core.tests import utils_fixtures
@@ -122,3 +126,85 @@ class ApilosSettingsViewTests(TestCase):
         self.assertFalse(new_user.is_superuser)
         self.assertTrue(new_user.is_bailleur())
         self.assertEqual('nicolas', new_user.creator.username)
+
+    def test_import_bailleur_users_upload_ok(self):
+        """
+        Superuser will login & upload a xlsx listing file
+        """
+        # login as superuser
+        self.client.post(
+            reverse("login"), {"username": "nicolas", "password": "12345"}
+        )
+
+        with open(os.path.join(os.path.dirname(__file__), './resources/listing_bailleur_ok.xlsx'), 'rb') as listing_file:
+            response = self.client.post(
+                reverse("settings:import_bailleur_users"),
+                {
+                    'file': listing_file,
+                    'Upload': True,
+                }
+            )
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_import_bailleur_users_submit_ok(self):
+        """
+        Superuser will login & submit a list of bailleurs users to create
+        """
+        # login as superuser
+        self.client.post(
+            reverse("login"), {"username": "nicolas", "password": "12345"}
+        )
+
+        response = self.client.post(reverse("settings:import_bailleur_users"))
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        bailleur = Bailleur.objects.get(nom='HLM')
+
+        response = self.client.post(
+            reverse("settings:import_bailleur_users"),
+            {
+                "form-TOTAL_FORMS": 1,
+                "form-INITIAL_FORMS": 1,
+                "form-0-first_name": "Jean",
+                "form-0-last_name": "Bailleur",
+                "form-0-bailleur": bailleur.id,
+                "form-0-email": "jean.bailleur2@apilos.com",
+                "form-1-first_name": "Jeanne",
+                "form-1-last_name": "Bailleur",
+                "form-1-bailleur": bailleur.id,
+                "form-1-email": "jeanne.bailleur2@apilos.com",
+                "form-2-first_name": "Chantal",
+                "form-2-last_name": "Bailleur",
+                "form-2-bailleur": bailleur.id,
+                "form-2-email": "chantal.bailleur2@apilos.com"
+
+            }
+        )
+        self.assertRedirects(response, reverse("settings:users"))
+
+    def test_import_bailleur_users_submit_nok(self):
+        """
+        Superuser will login & submit a list of bailleurs users to create
+        """
+        # login as superuser
+        self.client.post(
+            reverse("login"), {"username": "nicolas", "password": "12345"}
+        )
+
+        response = self.client.post(reverse("settings:import_bailleur_users"))
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        bailleur = Bailleur.objects.get(nom='HLM')
+
+        response = self.client.post(
+            reverse("settings:import_bailleur_users"),
+            {
+                "form-TOTAL_FORMS": 1,
+                "form-INITIAL_FORMS": 1,
+                "form-0-first_name": "Sabine",
+                "form-0-last_name": "L'autre",
+                "form-0-bailleur": bailleur.id,
+                "form-0-email": "sabine@apilos.com"
+            }
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
